# demo/generat/demogenapp-api-interna-client-exemplesecure-v1/api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: otae@fundaciobit.org
    name: Departament de Govern Digital a la Fundació Bit
    url: https://governdigital.fundaciobit.org
  description: Conjunt de Serveis REST de DemoGenApp per ser accedits emprant autenticació
  license:
    name: European Union Public Licence (EUPL v1.2)
    url: https://joinup.ec.europa.eu/sites/default/files/custom-page/attachment/eupl_v1.2_es.pdf
  title: API REST INTERNA de DemoGenApp - Exemple de Servei Securitzat
  version: 1.0-SNAPSHOT
externalDocs:
  description: Java Client (GovernIB Github)
  url: https://github.com/GovernIB/demogenapp/tree/demogenapp-1.0/demogenapp-api-interna-client-exemplesecure-v1
servers:
- url: /demogenappapi/interna
- url: http://localhost:8080/demogenappapi/interna
- url: https://dev.caib.es/demogenappapi/interna
- url: https://proves.caib.es/demogenappapi/interna
- url: https://se.caib.es/demogenappapi/interna
- url: https://www.caib.es/demogenappapi/interna
tags:
- description: Notificacions a l'APP de Carpeta (missatge a Mòbil)
  name: Notificacions
paths:
  /secure/exempleenum/list:
    get:
      operationId: list
      parameters:
      - description: Pàgina de la que es volen obtenir les dades. Comença per 1.
        example: 1
        explode: true
        in: query
        name: page
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - description: Quantitat d'elements a retornar
        example: 10
        explode: true
        in: query
        name: pagesize
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - description: Idioma en que s'han de retornar les dades(Només suportat 'ca'
          o 'es')
        examples:
          Català:
            description: Català
            value: ca
          Castellano:
            description: Castellano
            value: es
        explode: true
        in: query
        name: language
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TipusDocumentalsPaginacio'
          description: "EFIB: Retornades dades obertes correctament"
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestExceptionInfo'
          description: "EFIB: Paràmetres incorrectes"
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestExceptionInfo'
          description: "EFIB: No Autenticat"
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestExceptionInfo'
          description: "EFIB: No Autoritzat"
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestExceptionInfo'
          description: "EFIB: Error durant la consulta de les dades obertes"
      security:
      - BasicAuth: []
      summary: Retorna un llistat de tipus documentals
      tags:
      - Notificacions
      x-accepts:
      - application/json
  /secure/exempleenum/sendnotificationtomobile:
    get:
      operationId: sendNotificationToMobile
      parameters:
      - description: Exemple de parametre Llistat d'Strings
        explode: true
        in: query
        name: notificationParameters
        required: true
        schema:
          items:
            type: string
          type: array
        style: form
      - description: Idioma en que s'enviaran els missatges d'error
        example: ca
        explode: true
        in: query
        name: langError
        required: true
        schema:
          pattern: ^ca|es$
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendMessageResult'
          description: Operació realitzada correctament
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestExceptionInfo'
          description: Paràmetres incorrectes
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestExceptionInfo'
          description: No Autenticat
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestExceptionInfo'
          description: No Autoritzat
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestExceptionInfo'
          description: Error no controlat
      security:
      - BasicAuth: []
      summary: Envia un missatge al mòbil del ciutada a traves de l'App de Carpeta.
      tags:
      - Notificacions
      x-accepts:
      - application/json
components:
  schemas:
    TipusDocumentalsPaginacio:
      example:
        totalcount: 5
        data:
        - data
        - data
        pagesize: 6
        page: 0
        totalpages: 1
      properties:
        page:
          description: Número pàgina. Comença per 1.
          format: int32
          type: integer
        pagesize:
          description: Mida de pàgina
          format: int32
          type: integer
        totalpages:
          description: Número total de pàgines
          format: int32
          type: integer
        totalcount:
          description: Numero total d'elements
          format: int32
          type: integer
        data:
          description: Elements retornats
          items:
            description: Elements retornats
            type: string
          type: array
      required:
      - data
      - page
      - pagesize
      - totalcount
      - totalpages
      type: object
    RestExceptionInfo:
      description: Estructura de dades utilitzada per passar informació d'un error
      example:
        field: field
        errorMessage: errorMessage
        stackTraceCause: stackTraceCause
        errorCode: 5
        stackTrace: stackTrace
      properties:
        errorCode:
          description: Codi intern de l'error. Si l'Aplicació no gestiona codis d'error
            llavors val null.
          format: int32
          type: integer
        errorMessage:
          description: Missatge de l'error
          type: string
        stackTrace:
          description: Stacktrace de l'excepció
          type: string
        stackTraceCause:
          description: Stacktrace de l'excepció causant de l'error si n'hi hagués.
          type: string
        field:
          description: Indica el camp en que hi ha un error de validació.
          type: string
      required:
      - errorMessage
      type: object
    SendMessageResult:
      example:
        code: OK
        message: message
      properties:
        code:
          $ref: '#/components/schemas/SendMessageResultCode'
        message:
          description: Missatge si hi ha error
          type: string
      type: object
    SendMessageResultCode:
      description: Codi resultat de l'operació
      enum:
      - OK
      - UNKNOWN_ERROR
      - NOTIFICATION_CODE_DO_NOT_EXIST
      - NOTIFICATION_CODE_DISABLED
      - ENTITYCODE_DO_NOT_EXIST
      - ENTITY_DISABLED
      - PLUGIN_DISABLED
      - PLUGIN_ENTITY_DO_NOT_EXIST
      - PLUGIN_ENTITY_DISABLED
      - CITIZEN_DO_NOT_EXIST
      - ERROR_SENDING_NOTIFICATION
      type: string
  securitySchemes:
    BasicAuth:
      scheme: basic
      type: http